Added lamps, Fixed crashing on server
Added both regular and inverted hexorium lamps. In order for this to happen a lot of changes had to be made to automatic color handler registry. While doing this I also realized the current structure registry would cause issues on the dedicated server and reworked the block nad item base classes to instead have client only methods for all the parameters.