Added monoliths, XP drops, Other minor changes
As per request to make this version survival-friendly I need to add generation. For that I needed to add natural monoliths which I did. All they are missing is a special base model (in HexShapes), everything else is done. Also tweaked the xp drops of ores and added them to natural monoliths. Other minor changes such as renames